add img and more styling to UI

+++ modules/clearStore.mjs
@@ -1,7 +1,7 @@
-const clearStore = () => {
-    if(window.confirm("Are you sure you want to clear ALL data?\nthis action cannot be undone!")) {
-        window.localStorage.clear()
-    }
-}
+const alertMessage = 
+`Are you sure you want to delete ALL names from the scoreboard?
+this action cannot be undone!`
+
+const clearStore = () => window.confirm(alertMessage) && window.localStorage.clear() 
 
 export default clearStore
